# docs/product-kit/04_data_api/openapi.yaml
openapi: 3.1.0
info:
  title: Work Calendar API
  version: 1.0.0
  description: 工作日历小程序、APP 与管理后台统一接口草案。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /change-requests:
    get:
      tags: [Changes]
      summary: 查询改班申请
      parameters:
        - {$ref: '#/components/parameters/WorkspaceHeader'}
        - name: status
          in: query
          schema: {type: string, enum: [pending, approved, rejected, withdrawn, expired]}
        - name: from
          in: query
          schema: {type: string, format: date}
        - name: to
          in: query
          schema: {type: string, format: date}
        - name: page
          in: query
          schema: {type: integer, minimum: 1, default: 1}
      responses:
        '200':
          description: 申请列表
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ChangeRequest'}
    post:
      tags: [Changes]
      summary: 提交临时改班
      parameters:
        - {$ref: '#/components/parameters/WorkspaceHeader'}
        - {$ref: '#/components/parameters/IdempotencyHeader'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChangeRequestInput'}
      responses:
        '201':
          description: 已提交或已直接生效
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ChangeRequest'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceHeader:
      name: X-Workspace-Id
      in: header
      required: true
      schema: {type: string, format: uuid}
    IdempotencyHeader:
      name: Idempotency-Key
      in: header
      required: false
      schema: {type: string, minLength: 8, maxLength: 128}
...
  schemas:
...
    CustomShift:
      type: object
      required: [name, kind]
      properties:
        name: {type: string}
        kind: {type: string, enum: [work, rest, leave, training, travel, custom]}
        startTime: {type: [string, 'null']}
        endTime: {type: [string, 'null']}
        endsNextDay: {type: boolean}
        color: {type: string}
...
    ChangeRequestInput:
      type: object
      required: [scheduleInstanceId, requestedShift]
      properties:
        scheduleInstanceId: {type: string, format: uuid}
        requestedShift: {$ref: '#/components/schemas/CustomShift'}
        shiftTemplateId: {type: [string, 'null'], format: uuid}
        locationId: {type: [string, 'null'], format: uuid}
        reason: {type: [string, 'null'], maxLength: 500}
    ChangeRequest:
      type: object
      required: [id, status, originalSnapshot, requestedSnapshot, createdAt]
      properties:
        id: {type: string, format: uuid}
        scheduleInstanceId: {type: string, format: uuid}
        status: {type: string, enum: [pending, approved, rejected, withdrawn, expired]}
        originalSnapshot: {type: object, additionalProperties: true}
        requestedSnapshot: {type: object, additionalProperties: true}
        reason: {type: [string, 'null']}
        approvalNote: {type: [string, 'null']}
        createdAt: {type: string, format: date-time}
        decidedAt: {type: [string, 'null'], format: date-time}
...